feat(cli): dock-only dialog UX with flush accent edge; remove variant system
