+++ apps/cli/src/tui/dialog/index.ts
@@ -1,5 +1,6 @@
 // Cline's custom dialog system for the TUI: a drop-in replacement for the
-// former @opentui-ui/dialog dependency, with Cline-designed visual variants.
+// former @opentui-ui/dialog dependency. Dialogs render as full-width bottom
+// sheets over a lightly dimmed conversation.
 export { DialogManager, type FocusHost } from "./manager";
 export {
 	DialogProvider,
@@ -22,12 +23,6 @@ export type {
 	DialogSize,
 	DialogState,
 	DialogStyle,
-	DialogVariant,
 	PromptContext,
 	PromptOptions,
 } from "./types";
-export {
-	DEFAULT_DIALOG_VARIANT,
-	DIALOG_VARIANTS,
-	normalizeDialogVariant,
-} from "./variants";
